@@ -6,7 +6,7 @@
   "type": "module",
   "scripts": {
     "build": "rollup --config",
-    "serve": "serve dist -p 3004",
+    "dev": "serve dist -p 3004",
     "start": "node start.js"
   },
   "keywords": [],
@@ -15,8 +15,8 @@
   "dependencies": {
     "@types/history": "^5.0.0",
     "jquery": "^3.6.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.2",
+    "react-dom": "^18.0.2",
     "react-redux": "^7.2.6",
     "redux": "^4.2.1",
     "serve": "^11.3.2",
