@@ -1,9 +1,11 @@
 
 
-import React from 'react';
+import * as React from 'react';
 import ReactDom from 'react-dom';
+import ReactDomClient from 'react-dom/client';
 
 export default {
     'react': React,
-    'react-dom': ReactDom
+    'react-dom': ReactDom,
+    'react-dom/client': ReactDomClient,
 }
